+++ PyLesson_04/Cube2.py
@@ -1,6 +1,6 @@
 def calcSurf(side):
     global area
-    area = (6*(side*side))
+    area = float(6*(side*side))
 def display(are):
     print("The surface area of a cube whose sides are",side,"in length is","{:0.5f}".format(are),".")
 side = float(input("Please enter the length of the sides: "))
