+++ PyLesson_04/Rectangle2.py
@@ -1,6 +1,6 @@
 def calcPerim(le,wi):
    global perim
-   perim = int(le*2)+(wi*2)
+   perim = float((le*2)+(wi*2))
 
 def display(prm):
     print("Your rectangle is","{:0.5f}".format(prm),"sq ft around.")
